fix the mrc image size <2G problem.

+++ libEM/mrcio.cpp
@@ -123,7 +123,7 @@ bool MrcIO::is_valid(const void *first_block, off_t file_size)
 		(nx > 1 && nx < max_dim) && (ny > 0 && ny < max_dim) && (nz > 0 && nz < max_dim)) {
 #ifndef SPIDERMRC // Spider MRC files don't satisfy the following test
 		if (file_size > 0) {
-			off_t file_size1 = nx * ny * nz * get_mode_size(mrcmode) + sizeof(MrcHeader);
+			off_t file_size1 = (off_t)nx * (off_t)ny * (off_t)nz * (off_t)get_mode_size(mrcmode) + (off_t)sizeof(MrcHeader);
 			if (file_size == file_size1) {
 				return true;
 			}
